initial package refactoring, using serve.js. TODO: import media assets through google cloud buckets
