+++ .npmignore
@@ -1,0 +1,14 @@
+node_modules
+**/.Rhistory
+**/.Rproj.user
+**/.ipynb_checkpoints
+.idea
+.DS_Store.firebase/*.cache
+**/.DS_Store
+eslintrc.json
+.firebaserc
+.gitignore
+assets.json
+firebase.json
+rollup.config.js
+webpack.config.cjs
