+++ assets.json
@@ -1,0 +1,66 @@
+{
+  "default": {
+      "languageSpecific": {
+          "device": [],
+          "shared": []
+      },
+      "shared": [
+          "bird-end.mp3",
+          "bird-halfway.mp3",
+          "bird-intro.mp3",
+          "cat-end.mp3",
+          "cat-halfway.mp3",
+          "cat-intro.mp3",
+          "deer-end.mp3",
+          "deer-halfway.mp3",
+          "deer-intro.mp3",
+          "dog-end.mp3",
+          "dog-halfway.mp3",
+          "dog-intro.mp3",
+          "end-ns.mp3",
+          "feedback-0-correct.mp3",
+          "feedback-0-incorrect.mp3",
+          "feedback-1-correct.mp3",
+          "feedback-1-incorrect.mp3",
+          "feedback-2-correct.mp3",
+          "feedback-2-incorrect.mp3",
+          "feedback-3-correct.mp3",
+          "feedback-3-incorrect.mp3",
+          "grade-select.mp3",
+          "halfway-ns.mp3",
+          "introduction-ns.mp3",
+          "introduction.mp3",
+          "neutral-sound-2.wav",
+          "neutral-sound-sub.mp3",
+          "neutral-sound.mp3",
+          "post-practice-instruction.mp3",
+          "post-practice-reminder.mp3",
+          "practice-instruction.mp3",
+          "student-select.mp3",
+          "academy-background.png",
+          "bird-happy.gif",
+          "bird-talking.gif",
+          "bird.png",
+          "cat-happy.gif",
+          "cat-talking.gif",
+          "cat.png",
+          "classroom-background.png",
+          "deer-happy.gif",
+          "deer-talking.gif",
+          "deer.png",
+          "dog-happy.gif",
+          "dog-talking.gif",
+          "dog.png",
+          "false-flash-key.gif",
+          "false-key.png",
+          "grade-keyboard.png",
+          "grade-keys.png",
+          "practice-left.png",
+          "practice-right.png",
+          "roar-sre-keyboard.png",
+          "roar-sre-practice-keyboard.png",
+          "true-flash-key.gif",
+          "true-key.png"
+        ]
+  }
+}
