--- webpack.config.js
@@ -1,134 +1,0 @@
-const path = require('path');
-const webpack = require('webpack');
-// eslint-disable-next-line import/no-extraneous-dependencies
-const { merge } = require("webpack-merge");
-// eslint-disable-next-line import/no-extraneous-dependencies
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-
-const commonConfig = {
-  entry: {
-    index: path.resolve(__dirname, 'src', 'index.js'),
-  },
-  output: {
-    filename: '[name].[contenthash].bundle.js',
-    path: path.resolve(__dirname, 'dist'),
-    clean: {
-      keep: /\.git/,
-    },
-  },
-  optimization: {
-    moduleIds: 'deterministic',
-    runtimeChunk: 'single',
-    splitChunks: {
-      cacheGroups: {
-        vendor: {
-          test: /[\\/]node_modules[\\/]/,
-          name(module) {
-            // get the name. E.g. node_modules/packageName/not/this/part.js
-            // or node_modules/packageName
-            const packageName = module.context.match(/[\\/]node_modules[\\/](.*?)([\\/]|$)/)[1];
-
-            // npm package names are URL-safe, but some servers don't like @ symbols
-            return `npm.${packageName.replace('@', '')}`;
-          },
-          chunks: 'all',
-        },
-      },
-    },
-  },
-  module: {
-    rules: [
-      {
-        test: /\.css$/i,
-        use: ['style-loader', 'css-loader'],
-      },
-      {
-        test: /\.(png|svg|jpg|jpeg|gif)$/i,
-        type: 'asset/resource',
-        generator: {
-          filename: 'img/[name][ext]',
-        },
-      },
-      {
-        test: /\.mp3$/,
-        use: [
-          {
-            loader: 'file-loader',
-            options: {
-              name: '[name].[ext]',
-              outputPath: 'audio',
-            },
-          },
-        ],
-      },
-      {
-        test: /\.mp4$/,
-        use: [
-          {
-            loader: 'file-loader',
-            options: {
-              name: '[name].[ext]',
-              outputPath: 'video',
-            },
-          },
-        ],
-      },
-      {
-        test: /\.csv$/,
-        use: [
-          {
-            loader: 'file-loader',
-            options: {
-              name: '[name].[ext]',
-              outputPath: 'csv',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  experiments: {
-    topLevelAwait: true,
-  },
-};
-
-const productionConfig = {
-  mode: "production",
-};
-
-const developmentConfig = {
-  mode: "development",
-  devtool: "inline-source-map",
-  devServer: {
-    static: "./dist",
-  },
-};
-
-module.exports = (env, args) => {
-  const roarDbDoc = env.dbmode === "production" ? "production" : "development";
-
-  let merged;
-  switch (args.mode) {
-    case 'development':
-      merged = merge(commonConfig, developmentConfig);
-      break;
-    case 'production':
-      merged = merge(commonConfig, productionConfig);
-      break;
-    default:
-      throw new Error('No matching configuration was found!');
-  }
-
-  return merge(
-    merged,
-    {
-      plugins: [
-        new HtmlWebpackPlugin({
-          title: 'Roar Anb',
-        }),
-        new webpack.ids.HashedModuleIdsPlugin(), // so that file hashes don't change unexpectedly
-        new webpack.DefinePlugin({ ROAR_DB_DOC: JSON.stringify(roarDbDoc) }),
-      ],
-    },
-  );
-};
